Add CUDA Error Handling (with a preprocessor macro). This fixes #3
The preprocessor macro is kind-of hacky, but that's when you get when using a C (?) preprocessor in a Fortran code. If this wasn't CUDA Fortran, the macro would make a lot of problems: In many cases, it will exceed the column (/line) length in Fortran; both function call and program filename play a role. Since CUDA Fortran is (as of writing this) only available in PGI's Fortran compiler, the problem is not as pronounced. Apparently PGI Fortran allows lines with 256 columns.

I tried to strip all unnecessary characters from the line to keep it minimal.

*  "err /= cudaSuccess" has been modified to "e/=CUDA_SUCCESS" which gets preprocessed to "e/=0", which seems to be the value of cudaSucess for the time being. It saves at least 10 characters!!!1
*  The FORMAT string for the write(*,FORMAT) is a variable. I created an inline  module to hide it from the main program.
*  CUDACALL actually calls CUDACALL__. The latter has arguments for the error code and the format string; the former does not. CUDACALL() should be the entry point in 98.2% of all cases, while CUDACALL__() is provided if anyone wants to use their own error code (to do something with this afterwards) or format string.

Kernels are not yet error-checked. This would mean another macro.